renamed functions2.py to processImages.py
